change courser to get double instead of long change datatype for energy consumption

+++ app/src/main/java/de/htwg/co2footprint_tracker/MainActivity.java
@@ -184,7 +184,7 @@ public class MainActivity extends AppCompatActivity {
                 packet.setReceivedPacketsWifi(cursor.getLong(3));
                 packet.setReceivedPacketsMobile(cursor.getLong(4));
                 packet.setReceivedPacketsTotal(cursor.getLong(5));
-                packet.setEnergyConsumption(cursor.getLong(6));
+                packet.setEnergyConsumption(cursor.getDouble(6));
 
             }
         }
